feat: add attack square consideration to position evaluation

+++ src/movefinding/board.h
@@ -56,8 +56,8 @@ typedef struct Position_t
     ULL en_passant_bitboard;
     bool white_to_move;
     uint8_t num_children;
-    uint8_t half_move_count;
-    int16_t piece_value_diff;
+    uint16_t half_move_count;
+    int64_t piece_value_diff;
     PiecesOneColour_t pieces[2];
     struct Position_t* parent_position;
     struct Position_t* child_positions[MAX_CHILDREN];
